backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-environ==0.8.1
psycopg[binary]==3.3.3